+++ UVtools.Core/FileFormats/ChituboxFile.cs
@@ -1068,10 +1068,10 @@ namespace UVtools.Core.FileFormats
         public override FileExtension[] FileExtensions { get; } = {
             new(typeof(ChituboxFile), "photon", "Chitubox Photon"),
             new(typeof(ChituboxFile), "cbddlp", "Chitubox CBDDLP"),
-            new(typeof(ChituboxFile), "ctb", $"Chitubox CTBv{USED_VERSION}"),
+            new(typeof(ChituboxFile), "ctb", "Chitubox CTB"),
             //new(typeof(ChituboxFile), "v2.ctb", "Chitubox CTBv2"),
             //new(typeof(ChituboxFile), "v3.ctb", "Chitubox CTBv3"),
-            new(typeof(ChituboxFile), "v4.ctb", "Chitubox CTBv4", false),
+            //new(typeof(ChituboxFile), "v4.ctb", "Chitubox CTBv4", false, false),
             //new(typeof(ChituboxFile), "encrypted.ctb", "Chitubox encrypted CTB"),
         };
 
@@ -1205,6 +1205,8 @@ namespace UVtools.Core.FileFormats
         };
 
         public override uint[] AvailableVersions { get; } = { 1, 2, 3, 4 };
+
+        public override uint DefaultVersion => USED_VERSION;
 
         public override uint Version
         {